Small changes
Next step is to change the method getting wing surfaces info, making the aero surfaces contain the main information as all the coefficents are from data of an aircraft as a whole and not the individual wings. The childs of aerosurface will be used to send information about the area and also using the area the aerosurfaces will determine the ratios to split the forces among different wing surfaces.